[fix]엔티티 수정 (#66)
필드 타입 변경
`BigDecimal` -> `Long`으로 수정
반정규화 일부 추가
`@OneToMany`관계 추가

+++ src/main/java/org/kakaoshare/backend/domain/funding/dto/RegisterRequest.java
@@ -1,17 +1,17 @@
 package org.kakaoshare.backend.domain.funding.dto;
 
-import java.math.BigDecimal;
-import java.time.LocalDate;
 import lombok.Builder;
 import lombok.Getter;
 import org.kakaoshare.backend.domain.funding.entity.Funding;
 import org.kakaoshare.backend.domain.member.entity.Member;
 import org.kakaoshare.backend.domain.product.entity.Product;
 
+import java.time.LocalDate;
+
 @Getter
 @Builder
 public class RegisterRequest {
-    private BigDecimal goalAmount;
+    private Long goalAmount;
     private LocalDate expiredAt;
 
     public Funding toEntity(Member member, Product product){
